LPD-41116 Move cluster tests to client-extension-web folder
Since it is a specific case of that project.
