+++ modules/test/playwright/tests/client-extension-web/tests/cluster/cluster.spec.ts
@@ -5,10 +5,10 @@
 
 import {expect, mergeTests} from '@playwright/test';
 
-import {loginTest} from '../../fixtures/loginTest';
-import {liferayConfig} from '../../liferay.config';
-import getRandomString from '../../utils/getRandomString';
-import {editJSImportMapsPageTest} from '../client-extension-web/fixtures/editJSImportMapsExtensionPageTest';
+import {loginTest} from '../../../../fixtures/loginTest';
+import {liferayConfig} from '../../../../liferay.config';
+import getRandomString from '../../../../utils/getRandomString';
+import {editJSImportMapsPageTest} from '../../../client-extension-web/fixtures/editJSImportMapsExtensionPageTest';
 
 export const test = mergeTests(editJSImportMapsPageTest, loginTest());
