+++ modules/test/playwright/tests/client-extension-web/tests/cluster/test.properties
